tests/wpt/web-platform-tests/referrer-policy/generic/inheritance/iframe-inheritance-data.html

<html>
  <head>
    <title>Referrer Policy: iframes with data url uses no referrer</title>
    <link rel="help" href="https://www.w3.org/TR/referrer-policy/#referrer-policy-delivery-nested">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/common/security-features/resources/common.sub.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body onload="runTest()">
    <h1>Referrer Policy: iframes with data url uses no referrer</h1>
    <script>
      let test = async_test("iframes with data url uses no referrer");
      window.addEventListener("message", test.step_func((msg) => {
        assert_equals(msg.data.referrer, undefined);
        test.done();
      }));

      function runTest() {
        const BASE = location.protocol + "//www1." + location.hostname + ":" + location.port;
        const TEST_DATA =
            `data:text/html,<script src = "${BASE}/common/security-features/resources/common.sub.js"></` + `script>
            <script src = "${BASE}/referrer-policy/generic/referrer-policy-test-case.js"></` + `script>
            <script>
              var urlPath = "/referrer-policy/generic/subresource/xhr.py";
              var url = "${BASE}" + urlPath;
              requestViaXhr(url).then((msg) => {
                  parent.postMessage({referrer: msg.referrer}, "*")})
                .catch((e) => {
                    parent.postMessage({referrer: "FAILURE"}, "*");
                  });
            </` + "script>";

        let iframe = document.createElement("iframe");
        iframe.src = TEST_DATA;
        document.body.appendChild(iframe);
      }
    </script>
    <div id="log"></div>
  </body>
</html>
