third_party/blink/web_tests/external/wpt/referrer-policy/generic/inheritance/iframe-inheritance-data-expected.txt

This is a testharness.js-based test.
FAIL iframes with data url uses no referrer assert_equals: expected (undefined) undefined but got (string) "FAILURE"
Harness: the test ran to completion.
